# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Xdash Rest-API ",
    "description": "Xdash Rest-API "
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "securityDefinitions": {
    "Bearer": {
      "name": "Authorization",
      "description": "API Token",
      "type": "apiKey",
      "in": "header"
    }
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Register new user in system",
        "description": "Register new user in system",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "User that we want to create",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "New user is created",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "401": {
            "description": "UnAuthorized",
            "schema": {
              "$ref": "#/definitions/UnAuthorized"
            }
          }
        }
      }
    },
...
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "User Login",
        "description": "Login user in system",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "schema": {
              "required": [
                "email",
                "password"
              ],
              "properties": {
                "email": {
                  "type": "string",
                  "value": "user@gmail.com"
                },
                "password": {
                  "type": "string",
                  "value": "devpass"
                }
              }
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "User logged in successfully.",
            "schema": {
              "type": "object",
              "properties": {
                "token": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "UnAuthorized",
            "schema": {
              "$ref": "#/definitions/UnAuthorized"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "User": {
      "required": [
        "email",
        "username",
        "password",
        "xdashId",
        "tokens",
        "fname",
        "lname"
      ],
      "properties": {
        "email": {
          "type": "string",
          "uniqueItems": true,
          "default": "tsestUser@gmail.com"
        },
        "username": {
          "type": "string",
          "uniqueItems": true
        },
        "password": {
          "type": "string",
          "uniqueItems": true
        },
        "xdashId": {
          "type": "string",
          "uniqueItems": true
        },
        "fname": {
          "type": "string",
          "uniqueItems": true,
          "default": "test"
        },
        "lname": {
          "type": "string",
          "uniqueItems": true
        }
      }
    },
...
    "UnAuthorized": {
      "properties": {
        "msg": {
          "type": "array",
          "items": {
            "properties": {
              "location": {
                "type": "string"
              },
              "param": {
                "type": "string"
              },
              "msg": {
                "type": "string"
              }
            }
          }
        }
      }
    }
  }
}